# 04-Projects/api/openapi.yml
openapi: 3.0.0
paths:
...
  /caro/matches:
    post:
      operationId: MatchController_create
      summary: Create a new match (authentication required)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMatchDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      data:
                        $ref: '#/components/schemas/CreateMatchResponseDto'
        '404':
          description: Config not found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 404
                      data:
                        nullable: true
                        example: null
        '409':
          description: Player already in active match
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 409
                      data:
                        nullable: true
                        example: null
      tags:
        - Caro — Matches
      security:
        - bearer: []
  /caro/matches/{id}/join:
    post:
      operationId: MatchController_join
      summary: Join a public lobby match (authentication required)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      data:
                        $ref: '#/components/schemas/JoinMatchResponseDto'
        '404':
          description: Match not found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 404
                      data:
                        nullable: true
                        example: null
        '409':
          description: Match not open or player already active
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 409
                      data:
                        nullable: true
                        example: null
      tags:
        - Caro — Matches
      security:
        - bearer: []
  /caro/matches/{id}:
    delete:
      operationId: MatchController_cancel
      summary: Cancel a match (creator only, before it starts; authentication required)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      data:
                        $ref: '#/components/schemas/CancelOrLeaveResponseDto'
        '403':
          description: Not the match creator
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 403
                      data:
                        nullable: true
                        example: null
        '409':
          description: Match cannot be cancelled in current status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 409
                      data:
                        nullable: true
                        example: null
      tags:
        - Caro — Matches
      security:
        - bearer: []
    get:
      operationId: MatchController_getState
      summary: >-
        Get full match state including moves (no authentication required for
        public matches)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      data:
                        $ref: '#/components/schemas/MatchStateDto'
        '404':
          description: Match not found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 404
                      data:
                        nullable: true
                        example: null
      tags:
        - Caro — Matches
  /caro/matches/{id}/leave:
    post:
      operationId: MatchController_leave
      summary: Leave a match before it starts (second player; authentication required)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      data:
                        $ref: '#/components/schemas/CancelOrLeaveResponseDto'
        '403':
          description: Not a participant
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 403
                      data:
                        nullable: true
                        example: null
        '409':
          description: Match not in waiting_for_start status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 409
                      data:
                        nullable: true
                        example: null
      tags:
        - Caro — Matches
      security:
        - bearer: []
  /caro/matches/{id}/invite:
    post:
      operationId: MatchController_invite
      summary: Invite a friend to a private match (authentication required)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitePlayerDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      data:
                        $ref: '#/components/schemas/InviteResponseDto'
        '403':
          description: Not the match creator
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 403
                      data:
                        nullable: true
                        example: null
        '422':
          description: Not friends with target player
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 422
                      data:
                        nullable: true
                        example: null
      tags:
        - Caro — Matches
      security:
        - bearer: []
  /caro/matches/{id}/invitation/respond:
    put:
      operationId: MatchController_respondToInvitation
      summary: Accept or decline a match invitation (authentication required)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespondInvitationDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      data:
                        $ref: '#/components/schemas/InvitationRespondResponseDto'
        '404':
          description: Match not found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 404
                      data:
                        nullable: true
                        example: null
        '409':
          description: Player already in active match
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 409
                      data:
                        nullable: true
                        example: null
      tags:
        - Caro — Matches
      security:
        - bearer: []
  /caro/matches/{id}/moves:
    get:
      operationId: MatchController_getMoves
      summary: >-
        Get all moves for a match (no authentication required for public
        matches)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/MoveDto'
        '404':
          description: Match not found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 404
                      data:
                        nullable: true
                        example: null
      tags:
        - Caro — Matches
    post:
      operationId: GameplayController_move
      summary: Place a move on the board
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceMoveDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      data:
                        $ref: '#/components/schemas/PlaceMoveResponseDto'
        '403':
          description: Not your turn or not a participant
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 403
                      data:
                        nullable: true
                        example: null
        '409':
          description: Cell occupied or deadline expired
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseDto'
                  - properties:
                      statusCode:
                        example: 409
                      data:
                        nullable: true
                        example: null
      tags:
        - Caro — Gameplay
      security:
        - bearer: []
...
info:
  title: Game Hub API
  description: Account & Social Module
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ApiResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 200
          description: Internal status code (mirrors HTTP status)
        data:
          type: object
          description: Response payload, shape varies per endpoint
        message:
          type: string
          example: Success
          description: Human-readable message
      required:
        - statusCode
        - data
        - message
...
    CreateMatchResponseDto:
      type: object
      properties:
        id:
          type: string
        configId:
          type: string
        boardSize:
          type: string
        moveTimeSeconds:
          type: number
        visibility:
          type: string
        status:
          type: string
        creatorId:
          type: string
        createdAt:
          format: date-time
          type: string
      required:
        - id
        - configId
        - boardSize
        - moveTimeSeconds
        - visibility
        - status
        - creatorId
        - createdAt
    CreateMatchDto:
      type: object
      properties:
        configId:
          type: string
          description: Active game configuration UUID
        visibility:
          type: string
          enum:
            - public
            - private
      required:
        - configId
        - visibility
    JoinMatchResponseDto:
      type: object
      properties:
        matchId:
          type: string
        status:
          type: string
      required:
        - matchId
        - status
    CancelOrLeaveResponseDto:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
      required:
        - id
        - status
    InviteResponseDto:
      type: object
      properties:
        matchId:
          type: string
        invitedPlayerId:
          type: string
      required:
        - matchId
        - invitedPlayerId
    InvitePlayerDto:
      type: object
      properties:
        friendId:
          type: string
          description: UUID of the friend to invite
      required:
        - friendId
    InvitationRespondResponseDto:
      type: object
      properties:
        matchId:
          type: string
        status:
          type: string
      required:
        - matchId
        - status
    RespondInvitationDto:
      type: object
      properties:
        action:
          type: string
          enum:
            - accept
            - decline
      required:
        - action
    PlayerInMatchDto:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        elo:
          type: number
        winRate:
          type: number
      required:
        - id
        - username
        - elo
        - winRate
    MoveDto:
      type: object
      properties:
        playerId:
          type: string
        row:
          type: number
        col:
          type: number
        sequenceNumber:
          type: number
        placedAt:
          format: date-time
          type: string
      required:
        - playerId
        - row
        - col
        - sequenceNumber
        - placedAt
    MatchStateDto:
      type: object
      properties:
        id:
          type: string
        boardSize:
          type: string
        moveTimeSeconds:
          type: number
        visibility:
          type: string
        status:
          type: string
        creatorId:
          type: string
        playerX:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/PlayerInMatchDto'
        playerO:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/PlayerInMatchDto'
        currentTurnPlayerId:
          type: object
          nullable: true
        deadlineAt:
          type: object
          nullable: true
        moves:
          type: array
          items:
            $ref: '#/components/schemas/MoveDto'
        viewers:
          type: array
          items:
            type: string
        pendingDrawRequestFromId:
          type: object
          nullable: true
        result:
          type: object
          nullable: true
        winnerPlayerId:
          type: object
          nullable: true
        startedAt:
          type: object
          nullable: true
        endedAt:
          type: object
          nullable: true
        createdAt:
          format: date-time
          type: string
      required:
        - id
        - boardSize
        - moveTimeSeconds
        - visibility
        - status
        - creatorId
        - playerX
        - playerO
        - currentTurnPlayerId
        - deadlineAt
        - moves
        - viewers
        - pendingDrawRequestFromId
        - result
        - winnerPlayerId
        - startedAt
        - endedAt
        - createdAt
    PlaceMoveResponseDto:
      type: object
      properties:
        playerId:
          type: string
        row:
          type: number
        col:
          type: number
        sequenceNumber:
          type: number
        placedAt:
          format: date-time
          type: string
        isGameOver:
          type: boolean
        result:
          type: object
          nullable: true
        winnerPlayerId:
          type: object
          nullable: true
      required:
        - playerId
        - row
        - col
        - sequenceNumber
        - placedAt
        - isGameOver
        - result
        - winnerPlayerId
    PlaceMoveDto:
      type: object
      properties:
        row:
          type: number
          description: Row index (0-based)
        col:
          type: number
          description: Column index (0-based)
      required:
        - row
        - col
...